ollama-pydantic-ai/pyproject.toml
[project]
name = "ollama-pydantic-ai"
version = "0.1.0"
description = "Ollama Pydantic AI"
authors = "Denis Burakov"
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.68.2",
    "orjson>=3.10.16",
    "pydantic-ai>=0.0.43",
    "python-dotenv>=1.0.1",
    "retry>=0.9.2",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "numpy>=2.0.2",
    "ruff>=0.11.2",
]
